Added exit code check

+++ build_pack_test.sh
@@ -3,7 +3,12 @@
 cd "$(dirname "$0")"
 
 echo "Compiling..."
-msbuild ./Sonic4_ModLoader/Sonic4_ModLoader.sln /p:Configuration=Release
+msbuild ./Sonic4_ModLoader/Sonic4_ModLoader.sln /p:Configuration=Release -m
+
+EXIT_CODE="$?"
+if [ "$EXIT_CODE" != "0" ]; then
+    exit $EXIT_CODE
+fi
 
 echo "Removing old distribution package..."
 rm -rf "./dist"
